PP-7938 Move Package Coordinates
- As part of the migration from Bintray to Maven Central we must move our coordinates to match the domain we own

- This means that the coordinates of our java packages will now exist under `uk.gov.service.payments` instead of `uk.gov.pay`

- This moves the dependencies of pay-java-commons to maven central from bintray
