+++ src/main/java/uk/gov/pay/adminusers/filters/LoggingMDCRequestFilter.java
@@ -7,8 +7,8 @@ import javax.ws.rs.container.ContainerRequestFilter;
 import java.io.IOException;
 import java.util.Optional;
 
-import static uk.gov.pay.logging.LoggingKeys.SERVICE_EXTERNAL_ID;
-import static uk.gov.pay.logging.LoggingKeys.USER_EXTERNAL_ID;
+import static uk.gov.service.payments.logging.LoggingKeys.SERVICE_EXTERNAL_ID;
+import static uk.gov.service.payments.logging.LoggingKeys.USER_EXTERNAL_ID;
 
 public class LoggingMDCRequestFilter implements ContainerRequestFilter {
     @Override
