+++ src/main/java/uk/gov/pay/adminusers/filters/LoggingMDCResponseFilter.java
@@ -8,8 +8,8 @@ import javax.ws.rs.container.ContainerResponseFilter;
 import java.io.IOException;
 import java.util.List;
 
-import static uk.gov.pay.logging.LoggingKeys.SERVICE_EXTERNAL_ID;
-import static uk.gov.pay.logging.LoggingKeys.USER_EXTERNAL_ID;
+import static uk.gov.service.payments.logging.LoggingKeys.SERVICE_EXTERNAL_ID;
+import static uk.gov.service.payments.logging.LoggingKeys.USER_EXTERNAL_ID;
 
 public class LoggingMDCResponseFilter implements ContainerResponseFilter {
     @Override
