+++ src/main/java/uk/gov/pay/adminusers/model/ServiceName.java
@@ -1,7 +1,7 @@
 package uk.gov.pay.adminusers.model;
 
 import uk.gov.pay.adminusers.persistence.entity.service.ServiceNameEntity;
-import uk.gov.pay.commons.model.SupportedLanguage;
+import uk.gov.service.payments.commons.model.SupportedLanguage;
 
 import java.util.Collection;
 import java.util.Collections;
