+++ src/test/java/uk/gov/pay/adminusers/model/ServiceNameTest.java
@@ -2,7 +2,7 @@ package uk.gov.pay.adminusers.model;
 
 import org.junit.jupiter.api.Test;
 import uk.gov.pay.adminusers.persistence.entity.service.ServiceNameEntity;
-import uk.gov.pay.commons.model.SupportedLanguage;
+import uk.gov.service.payments.commons.model.SupportedLanguage;
 
 import java.util.List;
 import java.util.Map;
