+++ src/test/java/uk/gov/pay/adminusers/pact/ContractTestSuite.java
@@ -5,7 +5,7 @@ import junit.framework.JUnit4TestAdapter;
 import junit.framework.TestSuite;
 import org.junit.runner.RunWith;
 import org.junit.runners.AllTests;
-import uk.gov.pay.commons.testing.pact.provider.CreateTestSuite;
+import uk.gov.service.payments.commons.testing.pact.provider.CreateTestSuite;
 
 
 @RunWith(AllTests.class)
